third_party/blink/web_tests/virtual/single-import-map/wpt_internal/import-maps/parsing_parsing-schema-scope.json-expected.txt

This is a testharness.js-based test.
[FAIL] parsing
  assert_equals: expected "{\\"imports\\":{},\\"integrity\\":{},\\"scopes\\":{}}" but got "null"
[FAIL] parsing 1
  assert_equals: expected "{\\"imports\\":{},\\"integrity\\":{},\\"scopes\\":{}}" but got "null"
[FAIL] parsing 2
  assert_equals: expected "{\\"imports\\":{},\\"integrity\\":{},\\"scopes\\":{}}" but got "null"
[FAIL] parsing 3
  assert_equals: expected "{\\"imports\\":{},\\"integrity\\":{},\\"scopes\\":{}}" but got "null"
[FAIL] parsing 4
  assert_equals: expected "{\\"imports\\":{},\\"integrity\\":{},\\"scopes\\":{}}" but got "null"
Harness: the test ran to completion.
